example/requirements.txt
django >= 2
django-filter >= 2.4.0